# src/backend/docs/swagger.yaml
basePath: /api/v1/nlip
definitions:
  clip.Clip:
    properties:
      clipId:
        type: string
      content:
        type: string
      contentType:
        type: string
      createdAt:
        type: string
      creator:
        $ref: '#/definitions/clip.Creator'
      filePath:
        type: string
      id:
        type: string
      spaceId:
        type: string
      updatedAt:
        type: string
    type: object
  clip.ClipResponse:
    properties:
      clip:
        $ref: '#/definitions/clip.Clip'
    type: object
  clip.Creator:
    properties:
      id:
        type: string
      username:
        type: string
    type: object
  clip.ListClipsResponse:
    properties:
      clips:
        items:
          $ref: '#/definitions/clip.Clip'
        type: array
    type: object
  clip.UpdateClipRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  clip.UploadClipRequest:
    properties:
      content:
        type: string
      contentType:
        type: string
      creator:
        type: string
      spaceId:
        type: string
    type: object
...
info:
  contact: {}
  description: Nlip API
  title: Nlip API
  version: "1.0"
paths:
...
  /api/v1/nlip/clips:
    get:
      consumes:
      - application/json
      description: 获取当前用户的Clip列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/clip.ListClipsResponse'
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Clip列表
      tags:
      - 剪贴板
    post:
      consumes:
      - application/json
      description: 上传剪贴板内容
      parameters:
      - description: 上传Clip请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clip.UploadClipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            $ref: '#/definitions/clip.ClipResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 上传Clip
      tags:
      - 剪贴板
  /api/v1/nlip/clips/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的Clip
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "403":
          description: 无权限删除
          schema:
            type: string
        "404":
          description: Clip不存在
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除Clip
      tags:
      - 剪贴板
    get:
      consumes:
      - application/json
      description: 根据ID获取单个Clip的详细信息
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/clip.ClipResponse'
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: Clip不存在
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Clip详情
      tags:
      - 剪贴板
    put:
      consumes:
      - application/json
      description: 更新指定Clip的信息
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新Clip请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clip.UpdateClipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/clip.ClipResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "403":
          description: 无权限修改
          schema:
            type: string
        "404":
          description: Clip不存在
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新Clip
      tags:
      - 剪贴板
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"